@@ -21,16 +21,16 @@
     "source": "https://github.com/mimmi20/browscap-json-generator"
   },
   "require": {
-    "php": ">= 5.6.0",
-    "browscap/browscap-php": "dev-master",
-    "browscap/browscap": "dev-master#bb28055",
+    "php": "^7.1",
+    "browscap/browscap-php": "^3.1",
+    "browscap/browscap": "dev-master#c380824",
     "mimmi20/browscap-json-cache": "^1.0",
-    "symfony/console": "^2.0 || ^3.0"
+    "symfony/console": "^3.3"
   },
   "require-dev": {
-    "phpunit/phpunit": "^5.3",
-    "squizlabs/php_codesniffer": "^2.5",
-    "friendsofphp/php-cs-fixer": "^1.11"
+    "phpunit/phpunit": "^6.2",
+    "squizlabs/php_codesniffer": "^2.9",
+    "friendsofphp/php-cs-fixer": "^2.3"
   },
   "autoload": {
     "psr-4": {
