@@ -1,20 +1,14 @@
 <?php
 /**
- * Copyright (c) 1998-2014 Browser Capabilities Project
+ * This file is part of the browscap-json-generator package.
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
+ * Copyright (c) 2012-2017, Thomas Mueller <mimmi20@live.de>
  *
- * Refer to the LICENSE file distributed with this package.
- *
- * @category   BrowscapTest
- *
- * @copyright  1998-2014 Browser Capabilities Project
- * @license    MIT
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
  */
 
+declare(strict_types = 1);
 namespace BrowscapTest;
 
 use Browscap\Data\PropertyHolder;
@@ -29,7 +23,7 @@ use Monolog\Logger;
  *
  * @author     James Titcumb <james@asgrim.com>
  */
-class UserAgentsJsonTest extends \PHPUnit_Framework_TestCase
+class UserAgentsJsonTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * Test to make sure the preprocessed json file is created
